move key manager to listener argument
