+++ extension/src/background/index.ts
@@ -7,7 +7,7 @@ import {
 import { ExternalRequest, Response } from "@shared/api/types";
 import { buildStore } from "background/store";
 
-import { popupMessageListener } from "./messageListener/popupMessageListener";
+import { popupMessageListener } from "./messageListener/popupMessageListener2";
 import { freighterApiMessageListener } from "./messageListener/freighterApiMessageListener";
 import { SESSION_ALARM_NAME, clearSession } from "./helpers/session";
 import {
@@ -20,6 +20,13 @@ import {
   dataStorageAccess,
   browserLocalStorage,
 } from "./helpers/dataStorageAccess";
+import { ServiceMessageRequest } from "@shared/api/types/message-request";
+import {
+  BrowserStorageKeyStore,
+  KeyManager,
+  ScryptEncrypter,
+} from "@stellar/typescript-wallet-sdk-km";
+import { BrowserStorageConfigParams } from "@stellar/typescript-wallet-sdk-km/lib/Plugins/BrowserStorageFacade";
 
 export const initContentScriptMessageListener = () => {
   browser?.runtime?.onMessage?.addListener((message) => {
@@ -35,12 +42,26 @@ export const initContentScriptMessageListener = () => {
 export const initExtensionMessageListener = () => {
   browser?.runtime?.onMessage?.addListener(async (request, sender) => {
     const sessionStore = await buildStore();
+    const localStore = dataStorageAccess(browserLocalStorage);
+    const localKeyStore = new BrowserStorageKeyStore();
+    localKeyStore.configure({
+      storage: browserLocalStorage as BrowserStorageConfigParams["storage"],
+    });
+    const keyManager = new KeyManager({
+      keyStore: localKeyStore,
+    });
+    keyManager.registerEncrypter(ScryptEncrypter);
     // todo this is kinda ugly
     const req = request as ExternalRequest | Response;
     let res;
 
     if (Object.values(SERVICE_TYPES).includes(req.type as SERVICE_TYPES)) {
-      res = await popupMessageListener(req as Response, sessionStore);
+      res = await popupMessageListener(
+        req as ServiceMessageRequest,
+        sessionStore,
+        localStore,
+        keyManager,
+      );
     }
     if (
       Object.values(EXTERNAL_SERVICE_TYPES).includes(
